+++ src/pages/_app.tsx
@@ -1,5 +1,7 @@
 import { CssBaseline } from '@mui/material';
 import '../styles/global.css';
+import { LocalizationProvider } from '@mui/x-date-pickers';
+import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import { Provider } from 'react-redux';
 
 import Layout from '../components/Layout';
@@ -10,10 +12,12 @@ import type { AppProps } from 'next/app';
 const MyApp = ({ Component, pageProps }: AppProps): JSX.Element => {
   return (
     <Provider store={store}>
-      <CssBaseline />
-      <Layout>
-        <Component {...pageProps} />
-      </Layout>
+      <LocalizationProvider dateAdapter={AdapterDayjs}>
+        <CssBaseline />
+        <Layout>
+          <Component {...pageProps} />
+        </Layout>
+      </LocalizationProvider>
     </Provider>
   );
 };